+++ CHANGELOG.md
@@ -10,11 +10,13 @@
 
 ### Changed
 
+- Update Bitmovin's native Android SDK version to `3.72.0`
 - Update Bitmovin's native iOS SDK version to `3.64.0`
 
 ### Fixed
 
 - Android: Can't create a new Player with an existing NativeID (to bind to the same native Player)
+- iOS: Performance problem with large WebVTT thumbnail or subtitle tracks
 
 ## [0.23.0] (2024-05-08)