+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.24.0]
+
+### Added
+
+- `NetworkConfig` to configure the network request manipulation functionality
+  - `preprocessHttpRequest(type, request) => Promise<HttpRequest>` to pre-process an HTTP request before the request is executed by the player
+  - `preprocessHttpResponse(type, response) => Promise<HttpResponse>` to pre-process an HTTP response before it gets passed to the player
+
+### Changed
+
+- Update Bitmovin's native Android SDK version to `3.72.0`
+- Update Bitmovin's native iOS SDK version to `3.64.0`
+
+### Fixed
+
+- Android: Can't create a new Player with an existing NativeID (to bind to the same native Player)
+- iOS: Performance problem with large WebVTT thumbnail or subtitle tracks
+
 ## [0.23.0] (2024-05-08)
 
 ### Changed